Moved PlanarTerrain to a dedicated file
