+++ src/hippopt/robot_planning/utilities/__init__.py
@@ -2,5 +2,6 @@ from . import (
     foot_contact_state_plotter,
     humanoid_state_visualizer,
     interpolators,
+    planar_terrain,
     terrain_descriptor,
 )
